Split GitService unit and integration tests
Keep tests that mock IRepository and create a real Repository separate.
